requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-groq===0.2.3
unstructured==0.16.13
selenium==4.27.1
chromadb==0.6.3
streamlit==1.41.1
pandas==2.2.3
python-dotenv==1.0.1
openpyxl==3.1.5
ollama==0.4.6